post_ops/requirements.txt
PyJWT>=1.7.1
sseclient>=0.0.26
azure-storage-blob
azure-identity
python-dotenv
